Move the DnsMessageUtil to util package
