+++ codec-dns/src/main/java/io/netty/handler/codec/dns/DefaultDnsPtrRecord.java
@@ -17,6 +17,7 @@ package io.netty.handler.codec.dns;
 
 import static io.netty.util.internal.ObjectUtil.checkNotNull;
 
+import io.netty.handler.codec.dns.util.DnsMessageUtil;
 import io.netty.util.internal.StringUtil;
 import io.netty.util.internal.UnstableApi;
