+++ codec-dns/src/main/java/io/netty/handler/codec/dns/DefaultDnsQuery.java
@@ -15,6 +15,7 @@
  */
 package io.netty.handler.codec.dns;
 
+import io.netty.handler.codec.dns.util.DnsMessageUtil;
 import io.netty.util.internal.UnstableApi;
 
 /**
