+++ codec-dns/src/main/java/io/netty/handler/codec/dns/DefaultDnsRawRecord.java
@@ -16,6 +16,7 @@
 package io.netty.handler.codec.dns;
 
 import io.netty.buffer.ByteBuf;
+import io.netty.handler.codec.dns.util.DnsMessageUtil;
 import io.netty.util.internal.StringUtil;
 import io.netty.util.internal.UnstableApi;
